% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{fes2.hillsfile3d}
\alias{fes2.hillsfile3d}
\title{Calculate 3D free energy surface by conventional algorithm}
\usage{
\method{fes2}{hillsfile3d}(hills, imin = 1, imax = NULL, xlim = NULL,
  ylim = NULL, zlim = NULL, npoints = NULL)
}
\arguments{
\item{hills}{hillsfile3d object.}

\item{imin}{index of a hill from which summation starts (default 1).}

\item{imax}{index of a hill from which summation stops (default the rest of hills).}

\item{xlim}{numeric vector of length 2, giving the CV1 coordinates range.}

\item{ylim}{numeric vector of length 2, giving the CV2 coordinates range.}

\item{zlim}{numeric vector of length 2, giving the CV3 coordinates range.}

\item{npoints}{resolution of the free energy surface in number of points.}
}
\value{
fes object.
}
\description{
`fes2.hills3d` sums up hills using slow conventional algorithm. It can be used
as a reference or when hill widths are variable.
}
\examples{
tfes<-fes2(acealanme3d, imax=100)
}
