% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{head.hillsfile3d}
\alias{head.hillsfile3d}
\title{Print first n lines of hillsfile3d}
\usage{
\method{head}{hillsfile3d}(x, n = 10, ...)
}
\arguments{
\item{x}{hillsfile3d object.}

\item{n}{number of lines (default 10).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`head.hillsfile3d` prints first n lines of a hillsfile3d object.
}
\examples{
head(acealanme3d)
}
