% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{max.fes3d}
\alias{max.fes3d}
\title{Calculate maximum of 3D free energy surface}
\usage{
\method{max}{fes3d}(inputfes, na.rm = NULL, ...)
}
\arguments{
\item{inputfes}{fes3d object.}

\item{na.rm}{a logical indicating whether missing values should be
removed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`max.fes3d` calculates maximum of free energy in a fes3d object.
}
\examples{
tfes<-fes(acealanme3d, imax=5000)
max(tfes)
}
