% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{min.fes3d}
\alias{min.fes3d}
\title{Calculate minimum of 3D free energy surface}
\usage{
\method{min}{fes3d}(inputfes, na.rm = NULL, ...)
}
\arguments{
\item{inputfes}{fes3d object.}

\item{na.rm}{a logical indicating whether missing values should be
removed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`min.fes3d` calculates minimum of free energy in a fes3d object.
}
\examples{
tfes<-fes(acealanme3d, imax=5000)
min(tfes)
}
