% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{print.fes3d}
\alias{print.fes3d}
\title{Print minimum and maximum of 3D free energy surface}
\usage{
\method{print}{fes3d}(x, ...)
}
\arguments{
\item{x}{fes3d object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`print.fes3d` prints dimensionality, minimum and maximum of
free energy in a fes object
}
\examples{
tfes<-fes(acealanme3d, imax=5000)
tfes
}
