% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{print.hillsfile3d}
\alias{print.hillsfile3d}
\title{Print hillsfile3d}
\usage{
\method{print}{hillsfile3d}(x, ...)
}
\arguments{
\item{x}{hillsfile3d object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`print.hillsfile3d` prints dimensionality and size of a hillsfile object.
}
\examples{
acealanme3d
}
