% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{print.minima3d}
\alias{print.minima3d}
\title{Print minima3d object}
\usage{
\method{print}{minima3d}(x, ...)
}
\arguments{
\item{x}{minima object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`print.minima3d` prints 3D free energy minima (identifier, values of bins and collective variables and free energy).
}
\examples{
tfes<-fes(acealanme3d, imax=5000)
minima<-fesminima(tfes)
minima
}
