% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{summary.fes3d}
\alias{summary.fes3d}
\title{Print summary of 3D free energy surface}
\usage{
\method{summary}{fes3d}(object, ...)
}
\arguments{
\item{object}{fes3d object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`summary.fes3d` prints minimum and maximum of free energy
in a fes object.
}
\examples{
tfes<-fes(acealanme3d, imax=5000)
summary(tfes)
}
