% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{summary.hillsfile3d}
\alias{summary.hillsfile3d}
\title{Print summary for hillsfile3d}
\usage{
\method{summary}{hillsfile3d}(object, ...)
}
\arguments{
\item{object}{hillsfile3d object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`summary.hillsfile3d` prints dimensionality, size and collective variable ranges of a hillsfile3d object.
}
\examples{
summary(acealanme3d)
}
