% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{tail.hillsfile3d}
\alias{tail.hillsfile3d}
\title{Print last n lines of hillsfile3d}
\usage{
\method{tail}{hillsfile3d}(x, n = 10, ...)
}
\arguments{
\item{x}{hillsfile3d object.}

\item{n}{number of lines (default 10).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`tail.hillsfile3d` prints last n lines of a hillsfile3d object.
}
\examples{
tail(acealanme3d)
}
