% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{plot.hillsfile3d}
\alias{plot.hillsfile3d}
\title{Plot hillsfile3d object}
\usage{
\method{plot}{hillsfile3d}(x, xlab = "CV1", ylab = "CV2", zlab = "CV3",
  main = NULL, sub = NULL, col = "orange", ...)
}
\arguments{
\item{x}{hillsfile object.}

\item{xlab}{a title for the x axis: see 'title'.}

\item{ylab}{a title for the y axis: see 'title'.}

\item{zlab}{a title for the z axis: see 'title'.}

\item{main}{an overall title for the plot: see 'title'.}

\item{sub}{a sub title for the plot: see 'title'.}

\item{col}{color code or name, see 'par'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`plot.hillsfile3d` plots hillsfile object. It plots CV1 vs CV2 vs CV3.
}
\examples{
plot(acealanme3d)
}
